# Trie node encode