This PR optimizes `Leaf` and `Extension` encoding, making it aligned with the one proposed in Paprika's [RlpStream](https://github.com/NethermindEth/Paprika/blob/main/src/Paprika/RLP/RlpStream.cs) but without introducing a new entity. An appropriate `Rlp.*` methods are used. Additionally, a but in `Rlp.Encode` is fixed.

## Benchmarks

### Before

| Method           | Mean      | Error    | StdDev   | Gen0   | Allocated |
|----------------- |----------:|---------:|---------:|-------:|----------:|
| Encode_Extension |  30.94 ns | 0.368 ns | 0.327 ns | 0.0083 |     104 B |
| Encode_Branch    | 380.42 ns | 4.567 ns | 4.272 ns | 0.0100 |     128 B |
| Encode_Leaf      |  23.64 ns | 0.251 ns | 0.223 ns | 0.0063 |      80 B |

### After

| Method           | Mean      | Error    | StdDev   | Gen0   | Allocated |
|----------------- |----------:|---------:|---------:|-------:|----------:|
| Encode_Extension |  24.45 ns | 0.322 ns | 0.301 ns | 0.0051 |      64 B |
| Encode_Branch    | 378.58 ns | 2.524 ns | 2.361 ns | 0.0100 |     128 B |
| Encode_Leaf      |  17.46 ns | 0.165 ns | 0.146 ns | 0.0032 |      40 B |

## Changes

- removes `RlpStream` from `Leaf` and `Extension` encoding making them cheaper to call
- fixes a bug in position tracking in `RLP` class

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [x] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [ ] Yes
- [ ] No

#### If yes, did you write tests?

- [ ] Yes
- [x] No

## Documentation

#### Requires documentation update

- [ ] Yes
- [x] No